uint instead of size_t for buffers

+++ lib/core/include/core/gpu/mesh.hpp
@@ -14,10 +14,10 @@ namespace VTX::Core::Gpu
 		std::vector<Vec3f>			   vertices;
 		std::vector<Vec3f>			   normals;
 		std::vector<Util::Color::Rgba> colors;
-		std::vector<size_t>			   visibilities;
-		std::vector<size_t>			   selections;
-		std::vector<size_t>			   ids;
-		std::vector<size_t>			   indices;
+		std::vector<uint>			   visibilities;
+		std::vector<uint>			   selections;
+		std::vector<uint>			   ids;
+		std::vector<uint>			   indices;
 	};
 } // namespace VTX::Core::Gpu
